+++ scripts/build.clj
@@ -1,0 +1,13 @@
+(require '[cljs.build.api :as b])
+
+(println "Building ...")
+
+(let [start (System/nanoTime)]
+  (b/build
+   (b/inputs "test" "src")
+   {:main 'superstring.core-test
+    :output-to "out/tests.js"
+    :output-dir "out"
+    :target :nodejs
+    :verbose true})
+  (println "... done. Elapsed" (/ (- (System/nanoTime) start) 1e9) "seconds"))