+++ src/superstring/core.cljs
@@ -469,7 +469,7 @@
     (join ""  (conj (map capitalize (rest words)) (lower-case (first words))))))
 
 (defn pascal-case
-  "Lower case first char in s and use capitalization to separate words.
+  "Upper case first char in s and use capitalization to separate words.
 
   foo bar => FooBar
   camelCase => CamelCase