+++ CHANGELOG.md
@@ -9,8 +9,8 @@
 # 1.1.0
 ### New features
 
-* Add char-at, which wraps the native String/charAt
-* Add re-quote which quotes a string, for use in regular expressions.
+* Add `char-at`, which wraps the native String/charAt
+* Add `re-quote` which quotes a string, for use in regular expressions.
 
 ### Bugs fixed